web-clone-agent/requirements.txt
# 网页抓取和解析
requests==2.28.2
beautifulsoup4==4.12.2
selenium==4.9.1
webdriver-manager==3.8.6

# 自然语言处理和AI
langchain==0.0.267
langchain-openai==0.0.1
openai==0.27.8
python-dotenv==1.0.0

# 前端代码处理
jsbeautifier==1.14.7
cssutils==2.7.0
html5lib==1.1

# 文件操作和工具
tqdm==4.65.0
colorama==0.4.6 